Extract foreach element compiler fixture
Move ForeachElementProbe beside its sole owning test while preserving the compiler-produced source shape.
